% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/announced-prefixes.R
\name{get_announced_prefixes}
\alias{get_announced_prefixes}
\title{Announced Prefixes}
\usage{
get_announced_prefixes(asn, starttime = NULL, endtime = NULL,
  min_peers_seeing = 3)
}
\arguments{
\item{asn}{The Autonomous System Number for which to return prefixes.
Will auto-prefix with \code{AS} if just numeric.}

\item{starttime, endtime}{Start/end times for the query. When not \code{NULL} should be
an ISO-8601 or Unix timestamp. If \code{starttime} is \code{NULL} the query defaults to two weeks
before current date and time. If \code{endtime} is \code{NULL}, the query falls back to current date and time.}

\item{min_peers_seeing}{Minimum number of RIS peers seeing the prefix for it to be included in the results.
Excludes low-visibility/localized announcements. Default: \code{3}.}
}
\description{
This data call returns all announced prefixes for a given ASN.
The results can be restricted to a specific time period.
}
\examples{
\dontrun{
get_announced_prefixes("AS3333")
}
}
