% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country-resource-list.R
\name{get_country_resource_list}
\alias{get_country_resource_list}
\title{Country Resource List}
\usage{
get_country_resource_list(iso2c, time = NULL, v4_format = NULL)
}
\arguments{
\item{iso2c}{The country to find IP prefixes and AS numbers for. (2-digit ISO-3166
country code - See \url{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2} for details.)}

\item{time}{The time to query. By default (\code{NULL}), returns the latest available data.
(ISO-8601 or Unix timestamp)}

\item{v4_format}{format parameter; when not \code{NULL}, possible values: \code{""} or \code{"prefix"}.
\code{"prefix"} will return each entry in prefix notation, meaning that ranges are
converted to CIDR prefixes.}
}
\description{
This data call lists the Internet resources associated with a country, including
ASNs, IPv4 ranges and IPv4/6 CIDR prefixes.
}
\examples{
\dontrun{
get_country_resource_list("MX")
}
}
