% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dns-chain.R
\name{get_dns_chain}
\alias{get_dns_chain}
\title{DNS Chain}
\usage{
get_dns_chain(host_or_ip)
}
\arguments{
\item{host_or_ip}{Hostname or IP address (Ipv4 or IPv6)}
}
\description{
This data call returns the recursive chain of DNS forward (A/AAAA/CNAME) and
reverse (PTR) records starting form either a hostname or an IP address.
}
\examples{
\dontrun{
get_dns_chain("r-project.org")
}
}
