% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-info.R
\name{get_network_info}
\alias{get_network_info}
\title{Network Info}
\usage{
get_network_info(ip)
}
\arguments{
\item{ip}{Any IP address one wants to get network info for.}
}
\description{
This data call returns the containing prefix and announcing ASN of a given IP address.
}
\examples{
\dontrun{
get_network_info("140.78.90.50")
}
}
