% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse-dns-ip.R
\name{reverse_dns_ip}
\alias{reverse_dns_ip}
\title{Reverse DNS IP}
\usage{
reverse_dns_ip(ip)
}
\arguments{
\item{ip}{IP address for the query.}
}
\description{
Simple lookup for the reverse DNS info against a single IP address.
}
\examples{
\dontrun{
reverse_dns_ip("193.0.6.139")
}
}
