% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripestat-package.R
\docType{package}
\name{ripestat}
\alias{ripestat}
\alias{ripestat-package}
\title{Query and Retrieve Data from the 'RIPEstat' 'API'}
\description{
'RIPEstat' is a web-based interface that provides everything you ever
wanted to know about 'IP' address space, 'Autonomous System Numbers' ('ASNs'), and
related information for hostnames and countries in one place. Methods are provides to
query and retrieve data from the 'RIPEstat' 'API'.
}
\author{
Bob Rudis (bob@rud.is)
}
